Add static prop
